frontend: Refactor `MaximumPriceCalculationExists`
- Move logic that's used only in the component a bit down to simplify parent component
